requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
dnspython==2.1.0
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pymongo==4.0.1
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
todoist-python==8.1.3
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
Werkzeug==2.0.2
